added test for ActivityUtil.emitWithActivityName
